+++ contracts/test/TestMessageBridgeProver.sol
@@ -67,7 +67,7 @@ contract TestMessageBridgeProver is MessageBridgeProver {
      */
     function prove(
         address _sender,
-        uint256 _sourceChainId,
+        uint64 _sourceChainId,
         bytes calldata _encodedProofs,
         bytes calldata /* _data */
     ) external payable override {
@@ -92,7 +92,7 @@ contract TestMessageBridgeProver is MessageBridgeProver {
      * @dev Returns a fixed fee amount for testing
      */
     function fetchFee(
-        uint256 /* domainID */,
+        uint64 /* domainID */,
         bytes calldata /* _encodedProofs */,
         bytes calldata /* _data */
     ) public view override returns (uint256) {
@@ -104,7 +104,7 @@ contract TestMessageBridgeProver is MessageBridgeProver {
      * @dev Just tracks that dispatch was called
      */
     function _dispatchMessage(
-        uint256 /* domainID */,
+        uint64 /* domainID */,
         bytes calldata /* encodedProofs */,
         bytes calldata /* data */,
         uint256 /* fee */
